[Lens] Split apart config panel component to more manageable chunks (#63910)
* [Lens] Split apart config panel component to more manageable chunks

* Moving around and renaming SASS appropriately

* Remove layer limit
