+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/_chart_switch.scss
