+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/_config_panel.scss
@@ -1,0 +1,7 @@
+.lnsConfigPanel__addLayerBtn {
+  color: transparentize($euiColorMediumShade, .3);
+  // Remove EuiButton's default shadow to make button more subtle
+  // sass-lint:disable-block no-important
+  box-shadow: none !important;
+  border: 1px dashed currentColor;
+}
