+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/_dimension_popover.scss
@@ -1,0 +1,9 @@
+.lnsDimensionPopover {
+  line-height: 0;
+  flex-grow: 1;
+}
+
+.lnsDimensionPopover__trigger {
+  max-width: 100%;
+  display: block;
+}
