+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/_index.scss
@@ -1,0 +1,4 @@
+@import 'chart_switch';
+@import 'config_panel';
+@import 'dimension_popover';
+@import 'layer_panel';
