+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/layer_actions.ts
@@ -5,8 +5,8 @@
  */
 
 import _ from 'lodash';
-import { EditorFrameState } from './state_management';
-import { Datasource, Visualization } from '../../types';
+import { EditorFrameState } from '../state_management';
+import { Datasource, Visualization } from '../../../types';
 
 interface RemoveLayerOptions {
   trackUiEvent: (name: string) => void;
