+++ x-pack/plugins/lens/public/editor_frame_service/editor_frame/index.scss
@@ -1,8 +1,6 @@
-@import 'chart_switch';
-@import 'config_panel_wrapper';
+@import 'config_panel/index';
 @import 'data_panel_wrapper';
 @import 'expression_renderer';
 @import 'frame_layout';
 @import 'suggestion_panel';
 @import 'workspace_panel_wrapper';
-
