+++ x-pack/plugins/lens/public/types.ts
@@ -273,7 +273,7 @@ export type VisualizationLayerWidgetProps<T = unknown> = VisualizationConfigProp
   setState: (newState: T) => void;
 };
 
-type VisualizationDimensionGroupConfig = SharedDimensionProps & {
+export type VisualizationDimensionGroupConfig = SharedDimensionProps & {
   groupLabel: string;
 
   /** ID is passed back to visualization. For example, `x` */
@@ -368,55 +368,86 @@ export interface VisualizationType {
 }
 
 export interface Visualization<T = unknown, P = unknown> {
+  /** Plugin ID, such as "lnsXY" */
   id: string;
 
-  visualizationTypes: VisualizationType[];
+  /**
+   * Initialize is allowed to modify the state stored in memory. The initialize function
+   * is called with a previous state in two cases:
+   * - Loadingn from a saved visualization
+   * - When using suggestions, the suggested state is passed in
+   */
+  initialize: (frame: FramePublicAPI, state?: P) => T;
+  /**
+   * Can remove any state that should not be persisted to saved object, such as UI state
+   */
+  getPersistableState: (state: T) => P;
 
+  /**
+   * Visualizations must provide at least one type for the chart switcher,
+   * but can register multiple subtypes
+   */
+  visualizationTypes: VisualizationType[];
+  /**
+   * If the visualization has subtypes, update the subtype in state.
+   */
+  switchVisualizationType?: (visualizationTypeId: string, state: T) => T;
+  /** Description is displayed as the clickable text in the chart switcher */
+  getDescription: (state: T) => { icon?: IconType; label: string };
+
+  /** Frame needs to know which layers the visualization is currently using */
   getLayerIds: (state: T) => string[];
+  /** Reset button on each layer triggers this */
   clearLayer: (state: T, layerId: string) => T;
+  /** Optional, if the visualization supports multiple layers */
   removeLayer?: (state: T, layerId: string) => T;
+  /** Track added layers in internal state */
   appendLayer?: (state: T, layerId: string) => T;
 
-  // Layer context menu is used by visualizations for styling the entire layer
-  // For example, the XY visualization uses this to have multiple chart types
-  getLayerContextMenuIcon?: (opts: { state: T; layerId: string }) => IconType | undefined;
-  renderLayerContextMenu?: (domElement: Element, props: VisualizationLayerWidgetProps<T>) => void;
-
+  /**
+   * For consistency across different visualizations, the dimension configuration UI is standardized
+   */
   getConfiguration: (
     props: VisualizationConfigProps<T>
   ) => { groups: VisualizationDimensionGroupConfig[] };
 
-  getDescription: (
-    state: T
-  ) => {
-    icon?: IconType;
-    label: string;
-  };
-
-  switchVisualizationType?: (visualizationTypeId: string, state: T) => T;
-
-  // For initializing from saved object
-  initialize: (frame: FramePublicAPI, state?: P) => T;
-
-  getPersistableState: (state: T) => P;
-
-  // Actions triggered by the frame which tell the datasource that a dimension is being changed
-  setDimension: (
-    props: VisualizationDimensionChangeProps<T> & {
-      groupId: string;
-    }
-  ) => T;
-  removeDimension: (props: VisualizationDimensionChangeProps<T>) => T;
-
-  toExpression: (state: T, frame: FramePublicAPI) => Ast | string | null;
+  /**
+   * Popover contents that open when the user clicks the contextMenuIcon. This can be used
+   * for extra configurability, such as for styling the legend or axis
+   */
+  renderLayerContextMenu?: (domElement: Element, props: VisualizationLayerWidgetProps<T>) => void;
+  /**
+   * Visualizations can provide a custom icon which will open a layer-specific popover
+   * If no icon is provided, gear icon is default
+   */
+  getLayerContextMenuIcon?: (opts: { state: T; layerId: string }) => IconType | undefined;
 
   /**
-   * Epression to render a preview version of the chart in very constraint space.
+   * The frame is telling the visualization to update or set a dimension based on user interaction
+   * groupId is coming from the groupId provided in getConfiguration
+   */
+  setDimension: (props: VisualizationDimensionChangeProps<T> & { groupId: string }) => T;
+  /**
+   * The frame is telling the visualization to remove a dimension. The visualization needs to
+   * look at its internal state to determine which dimension is being affected.
+   */
+  removeDimension: (props: VisualizationDimensionChangeProps<T>) => T;
+
+  /**
+   * The frame will call this function on all visualizations at different times. The
+   * main use cases where visualization suggestions are requested are:
+   * - When dragging a field
+   * - When opening the chart switcher
+   * If the state is provided when requesting suggestions, the visualization is active.
+   * Most visualizations will apply stricter filtering to suggestions when they are active,
+   * because suggestions have the potential to remove the users's work in progress.
+   */
+  getSuggestions: (context: SuggestionRequest<T>) => Array<VisualizationSuggestion<T>>;
+
+  toExpression: (state: T, frame: FramePublicAPI) => Ast | string | null;
+  /**
+   * Expression to render a preview version of the chart in very constrained space.
    * If there is no expression provided, the preview icon is used.
    */
   toPreviewExpression?: (state: T, frame: FramePublicAPI) => Ast | string | null;
-
-  // The frame will call this function on all visualizations when the table changes, or when
-  // rendering additional ways of using the data
-  getSuggestions: (context: SuggestionRequest<T>) => Array<VisualizationSuggestion<T>>;
 }
